Fix gcc sign-compare warning caused by r9819
Fix this build error when using "scons debug" with GCC 4.4.3:

g++ -o build/debug/editor/editloop.o -c -fsigned-char -g -Wsign-compare -Wall -Werror -DHAVE_STRCASESTR -DHAVE_STRNLEN -DHAVE_X -D_GNU_SOURCE=1 -D_REENTRANT -DDEBUG -I/usr/local/include -I/usr/include/lua5.1 -I/usr/include -I/usr/include/SDL -Iengine/include -Iengine/guichan/include engine/editor/editloop.cpp
cc1plus: warnings being treated as errors
engine/editor/editloop.cpp: In function 'void EditorMainLoop()':
engine/editor/editloop.cpp:2088: error: comparison between signed and unsigned integer expressions

"make.py debug" does not (yet?) add -Wsign-compare, so the warning did
not occur with that.

+++ engine/editor/editloop.cpp
@@ -2085,7 +2085,7 @@ static void EditorMainLoop()
 	editorUnitSlider->addActionListener(editorUnitSliderListener);
 
 	editorUnitListModel = new StringListModel();
-	for (int i = 0; i < sizeof(unitFilters) / sizeof(*unitFilters); ++i)
+	for (size_t i = 0; i < sizeof(unitFilters) / sizeof(*unitFilters); ++i)
 	{
 		editorUnitListModel->add(unitFilters[i].name);
 	}
